refactor: split oversized git models and group mode tests
Keep the max-lines ratchet green by moving history models into a part
file and isolating source control group-mode widget tests.
