+++ test/widget/workspace_git_diff_panel_test.dart
@@ -1395,91 +1395,6 @@ void main() {
     );
   });
 
-  testWidgets('groups are ordered as staged unstaged and untracked', (
-    tester,
-  ) async {
-    final backend = FakeGitBackend()
-      ..gitStatusResult = const GitStatusResult(
-        entries: <GitChangeEntry>[
-          GitChangeEntry(
-            path: 'lib/new.dart',
-            area: GitChangeArea.untracked,
-            status: GitChangeStatus.untracked,
-          ),
-          GitChangeEntry(
-            path: 'lib/dirty.dart',
-            area: GitChangeArea.unstaged,
-            status: GitChangeStatus.modified,
-          ),
-          GitChangeEntry(
-            path: 'lib/staged.dart',
-            area: GitChangeArea.staged,
-            status: GitChangeStatus.modified,
-          ),
-        ],
-      );
-
-    await _pumpPanel(tester, backend: backend);
-    await tester.pumpAndSettle();
-
-    expect(
-      tester.getTopLeft(find.text('Staged')).dy,
-      lessThan(tester.getTopLeft(find.text('Unstaged')).dy),
-    );
-    expect(
-      tester.getTopLeft(find.text('Unstaged')).dy,
-      lessThan(tester.getTopLeft(find.text('Untracked')).dy),
-    );
-  });
-
-  testWidgets(
-    'unified group mode shows one changes section with area markers',
-    (tester) async {
-      final backend = FakeGitBackend()
-        ..gitStatusResult = const GitStatusResult(
-          entries: <GitChangeEntry>[
-            GitChangeEntry(
-              path: 'lib/new.dart',
-              area: GitChangeArea.untracked,
-              status: GitChangeStatus.untracked,
-            ),
-            GitChangeEntry(
-              path: 'lib/dirty.dart',
-              area: GitChangeArea.unstaged,
-              status: GitChangeStatus.modified,
-            ),
-            GitChangeEntry(
-              path: 'lib/staged.dart',
-              area: GitChangeArea.staged,
-              status: GitChangeStatus.modified,
-            ),
-            GitChangeEntry(
-              path: 'lib/dirty.dart',
-              area: GitChangeArea.staged,
-              status: GitChangeStatus.modified,
-            ),
-          ],
-        );
-
-      await _pumpPanel(
-        tester,
-        backend: backend,
-        groupMode: GitDiffGroupMode.unified,
-      );
-      await tester.pumpAndSettle();
-
-      expect(find.text('Changes'), findsOneWidget);
-      expect(find.text('Staged'), findsNothing);
-      expect(find.text('Unstaged'), findsNothing);
-      expect(find.text('Untracked'), findsNothing);
-      expect(find.text('lib/new.dart'), findsOneWidget);
-      expect(find.text('lib/staged.dart'), findsOneWidget);
-      // Same path appears once staged and once unstaged.
-      expect(find.text('lib/dirty.dart'), findsNWidgets(2));
-      expect(find.text('S'), findsNWidgets(2));
-    },
-  );
-
   testWidgets('filter narrows visible files without flattening groups', (
     tester,
   ) async {
